+++ src/main/java/tornadofx/Controls.kt
@@ -462,7 +462,7 @@ fun TextInputControl.stripNonInteger() = textProperty().mutateOnChange { it?.rep
  */
 fun TextInputControl.stripNonNumeric(vararg allowedChars: String = arrayOf(".", ",")) = textProperty().mutateOnChange { it?.replace(Regex("[^0-9${allowedChars.joinToString("")}]"), "") }
 
-fun ChoiceBox<*>.action(op: ActionEvent.() -> Unit) = setOnAction(op)
-fun ButtonBase.action(op: ActionEvent.() -> Unit) = setOnAction(op)
-fun TextField.action(op: ActionEvent.() -> Unit) = setOnAction(op)
-fun MenuItem.action(op: ActionEvent.() -> Unit) = setOnAction(op)
+fun ChoiceBox<*>.action(op: () -> Unit) = setOnAction { op() }
+fun ButtonBase.action(op: () -> Unit) = setOnAction { op() }
+fun TextField.action(op: () -> Unit) = setOnAction { op() }
+fun MenuItem.action(op: () -> Unit) = setOnAction { op() }